requirements.txt
django>2.2,<3.0
django-extensions>=3.1
graphene-django~=2.15.0
jwt>=1.1.0
django-hurricane==0.5.0
docker~=4.1.0
pyyaml~=5.3.1
gitpython~=3.1.9
python-keycloak~=0.24.0
python-keycloak-client~=0.2.3